+++ tests/Integration/Services/Stubs/InstrumentedAssignmentDistributor.php
@@ -45,7 +45,7 @@ readonly class InstrumentedAssignmentDistributor extends AssignmentDistributor
     {
         $this->assignGroupRuns->push($run);
 
-        return [0, 0];
+        return [0, 0, 0];
     }
 
     public function buildGroups(Collection $poolVideos): Collection